# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AtomQuest Goal Setting & Tracking Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ─── USERS ──────────────────────────────────────────────
...
  /analytics/goal-completion:
    get:
      operationId: getGoalCompletion
      tags: [analytics]
      summary: Goal completion rates by department/quarter
      parameters:
        - name: quarter
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Goal completion data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompletionStat"
  /analytics/department-performance:
    get:
      operationId: getDepartmentPerformance
      tags: [analytics]
      summary: Department-level performance overview
      parameters:
        - name: quarter
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Department performance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeptPerformance"
  /analytics/employee-scores:
    get:
      operationId: getEmployeeScores
      tags: [analytics]
      summary: Employee weighted scores and rankings
      parameters:
        - name: quarter
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: managerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Employee scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeScore"
...
  /analytics/quarter-trend:
    get:
      operationId: getQuarterTrend
      tags: [analytics]
      summary: Quarter-on-quarter performance trend
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: department
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"
  /analytics/uom-breakdown:
    get:
      operationId: getUomBreakdown
      tags: [analytics]
      summary: Goals breakdown by unit of measurement
      parameters:
        - name: quarter
          in: query
          schema:
            type: string
      responses:
        "200":
          description: UoM breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UomStat"

  # ─── REPORTS ────────────────────────────────────────────
  /reports/export:
    get:
      operationId: exportReport
      tags: [analytics]
      summary: Export goals report as CSV
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [csv]
        - name: quarter
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: managerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CSV export
          content:
            text/csv:
              schema:
                type: string

  # ─── ADMIN ──────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── User ────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, role, department, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [employee, manager, admin]
        department:
          type: string
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    CompletionStat:
      type: object
      required: [label, total, completed, percentage]
      properties:
        label:
          type: string
        total:
          type: integer
        completed:
          type: integer
        percentage:
          type: number
    DeptPerformance:
      type: object
      required: [department, avgScore, totalEmployees, totalGoals]
      properties:
        department:
          type: string
        avgScore:
          type: number
        totalEmployees:
          type: integer
        totalGoals:
          type: integer
        completedGoals:
          type: integer
    EmployeeScore:
      type: object
      required: [employeeId, employeeName, department, weightedScore, totalGoals]
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        department:
          type: string
        managerName:
          type: ["string", "null"]
        weightedScore:
          type: number
        totalGoals:
          type: integer
        completedGoals:
          type: integer
        quarter:
          type: ["string", "null"]
...
    TrendPoint:
      type: object
      required: [quarter, avgScore]
      properties:
        quarter:
          type: string
        avgScore:
          type: number
        completed:
          type: integer
        total:
          type: integer
    UomStat:
      type: object
      required: [uom, count, avgScore]
      properties:
        uom:
          type: string
        count:
          type: integer
        avgScore:
          type: number

    # ── Department ──────────────────────────────────────
...