# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AtomQuest Goal Setting & Tracking Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  quarter: string;
  avgScore: number;
  completed?: number;
  total?: number;
}