# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AtomQuest Goal Setting & Tracking Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  department: string;
  /** @nullable */
  managerId?: number | null;
  /** @nullable */
  managerName?: string | null;
  createdAt: Date;
}